cli: reject duplicate project JSON fields (audit 15)
Share the existing unique-key validator and add schema-aware case folding for typed lock fields, preserving exact map and plugin-config keys. Bundle duplicate-field regressions.

Proof: project and registry suites plus typed JSON regressions pass. Docs: audit-fixes-2026-09.md.
