+++ cli/manager/internal/registry/oauth.go
@@ -1,21 +1,19 @@
 package registry
 
 import (
-	"bytes"
 	"context"
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
 	"net/http"
 	"net/url"
 	"os"
 	"os/exec"
 	"runtime"
 	"strings"
-	"unicode"
 
 	"github.com/wago-org/wago/cli/internal/automation"
+	"github.com/wago-org/wago/internal/jsonstrict"
 )
 
 var oauthResponseMaximum int64 = 128 << 10
@@ -58,152 +56,12 @@ func unmarshalUniqueJSON(data []byte, output any) error {
 	return nil
 }
 
-// ValidateUniqueJSON validates one JSON value and rejects exactly repeated
-// object members. JSON map keys remain case-sensitive.
-func ValidateUniqueJSON(data []byte) error {
-	return validateUniqueJSON(data, false, nil)
-}
+// ValidateUniqueJSON rejects exactly repeated JSON object keys.
+func ValidateUniqueJSON(data []byte) error { return jsonstrict.ValidateUniqueJSON(data) }
 
-// ValidateUniqueFoldedJSON additionally treats case-folded object member names
-// as duplicates, matching encoding/json's struct-field lookup. Values of fields
-// in exactSubtrees use ordinary case-sensitive JSON object semantics.
+// ValidateUniqueFoldedJSON also rejects case-folded keys outside exact subtrees.
 func ValidateUniqueFoldedJSON(data []byte, exactSubtrees ...string) error {
-	var exact map[string]struct{}
-	if len(exactSubtrees) > 0 {
-		exact = make(map[string]struct{}, len(exactSubtrees))
-	}
-	for _, field := range exactSubtrees {
-		exact[foldJSONName(field)] = struct{}{}
-	}
-	return validateUniqueJSON(data, true, exact)
-}
-
-func validateUniqueJSON(data []byte, foldNames bool, exactSubtrees map[string]struct{}) error {
-	decoder := json.NewDecoder(bytes.NewReader(data))
-	decoder.UseNumber()
-	var frames []uniqueJSONFrame
-	rootStarted := false
-	rootComplete := false
-	for {
-		token, err := decoder.Token()
-		if errors.Is(err, io.EOF) {
-			if !rootComplete {
-				return errors.New("JSON response is incomplete")
-			}
-			return nil
-		}
-		if err != nil {
-			return errors.New("JSON response is invalid")
-		}
-		if rootComplete {
-			return errors.New("JSON response contains multiple values")
-		}
-
-		if delimiter, ok := token.(json.Delim); ok {
-			switch delimiter {
-			case '{', '[':
-				if err := beginUniqueJSONValue(frames, &rootStarted); err != nil {
-					return err
-				}
-				childFoldNames := foldNames
-				if len(frames) > 0 {
-					childFoldNames = frames[len(frames)-1].valueFoldNames
-				}
-				frames = append(frames, uniqueJSONFrame{
-					object: delimiter == '{', wantKey: delimiter == '{',
-					foldNames: childFoldNames, valueFoldNames: childFoldNames,
-				})
-			case '}', ']':
-				if len(frames) == 0 {
-					return errors.New("JSON response contains an unexpected closing delimiter")
-				}
-				frame := frames[len(frames)-1]
-				if frame.object != (delimiter == '}') || frame.object && !frame.wantKey {
-					return errors.New("JSON response contains an invalid closing delimiter")
-				}
-				frames = frames[:len(frames)-1]
-				completeUniqueJSONValue(frames, &rootComplete)
-			default:
-				return errors.New("JSON response contains an unexpected delimiter")
-			}
-			continue
-		}
-
-		if len(frames) > 0 && frames[len(frames)-1].object && frames[len(frames)-1].wantKey {
-			key, ok := token.(string)
-			if !ok {
-				return errors.New("JSON response contains a non-string object key")
-			}
-			frame := &frames[len(frames)-1]
-			if frame.members == nil {
-				frame.members = map[string]struct{}{}
-			}
-			canonicalKey := key
-			if frame.foldNames {
-				canonicalKey = foldJSONName(key)
-			}
-			if _, exists := frame.members[canonicalKey]; exists {
-				return errors.New("JSON response contains a duplicate object field")
-			}
-			frame.members[canonicalKey] = struct{}{}
-			frame.valueFoldNames = frame.foldNames
-			if _, exact := exactSubtrees[foldJSONName(key)]; exact {
-				frame.valueFoldNames = false
-			}
-			frame.wantKey = false
-			continue
-		}
-
-		if err := beginUniqueJSONValue(frames, &rootStarted); err != nil {
-			return err
-		}
-		completeUniqueJSONValue(frames, &rootComplete)
-	}
-}
-
-type uniqueJSONFrame struct {
-	object         bool
-	wantKey        bool
-	foldNames      bool
-	valueFoldNames bool
-	members        map[string]struct{}
-}
-
-func beginUniqueJSONValue(frames []uniqueJSONFrame, rootStarted *bool) error {
-	if len(frames) == 0 {
-		if *rootStarted {
-			return errors.New("JSON response contains multiple values")
-		}
-		*rootStarted = true
-		return nil
-	}
-	if frame := frames[len(frames)-1]; frame.object && frame.wantKey {
-		return errors.New("JSON response contains an object value without a key")
-	}
-	return nil
-}
-
-func completeUniqueJSONValue(frames []uniqueJSONFrame, rootComplete *bool) {
-	if len(frames) == 0 {
-		*rootComplete = true
-		return
-	}
-	frame := &frames[len(frames)-1]
-	if frame.object {
-		frame.wantKey = true
-	}
-}
-
-func foldJSONName(name string) string {
-	return strings.Map(func(character rune) rune {
-		for {
-			next := unicode.SimpleFold(character)
-			if next <= character {
-				return next
-			}
-			character = next
-		}
-	}, name)
+	return jsonstrict.ValidateUniqueFoldedJSON(data, exactSubtrees...)
 }
 
 func OpenBrowser(target string) error {
